add Fortran subroutine CLASS + routines it needs (except REPORT)
The philosophy of Twinspan is very anti-R: it goes on and prints
while it goes, but hardly returns anything (except with IPUNCH).
I should change it to work silently and return the results for
further analysis. Therefore REPORT was left out, and some other
subroutines may go as well. At least the WRITE statements must be
removed later.