do not use I2CHAR function or printable character strings
We do not want to print anything from the Fortran code, but the
original code builds printable strings of indicator species with
their signs. These places are no changed to handle integer indices
of species with negative indicators given negative integers.

For some reason, when compiled within R package, function I2CHAR
is not found, although it seemed to be OK with the self-standing
twinspan.

Except mess that should be taken care once we get that far in
the function.