fix segfault with deep divisions
The original Fortran code hardcoded tot(511) to store the totals
of classes. However, the class number could be higher with >=9
levels of division, and this caused segfault. I don't know why
the self-standing TWINSPAN did not segfault, but it could not
produce species classification with this depth but claimed that
there are no species:
DIVISION    1  (N=   0)         I.E. GROUP *
DIVISION FAILS - THERE ARE TOO FEW ITEMS