R CMD check warned on missing documentation of eigenvals
eigenvals was imported from vegan and had to be exported so that
it could be used withoug loading vegan. However, this also means
that we need to document the method so that ?eigenvals could be
found when vegan is not loaded. Unfortunately, this makes things
messier when vegan is loaded & attached.