change way division heights are found with eigenheight=TRUE
We used have the height of mother division as height, and use
height 1 for the top division. Now we use division height. Now
terminal groups (which were not divided) have no height. An arbitrary
height of n/(n-1) proportion of mother division height is used for
terminal group of size n.