warn when an id is larger than integer maximum
With ids doubling with depth, integer maximum 2^31-1 is reached
at depth 30. For reliable classification and cutting the difference
between left and right ids should be exactly 1. This seems often
work with larger numbers when R uses double precision floats, but
it will break at some point (and I don't know which point). So i
warn.

Depth > 30 is most likely to occur in cluster methods that are
prone to chaining (joining single leaves to existing clusters),
in other words in single linkage clustering.