tree fails if a child is more heterogeneous than her parent
The simple method of re-arranging by homogeneity fails with tree
reversal (kid above mother): that will define mother before her
offspring and tree is mangled. Quick tests indicated that this is
not too common (I got this  with the mite data of vegan). Alternative
was to revert 820e8ba9eac, but since this seems to mostly work, we
stop only with reversals. Building tree from the bottom and only allowing
adding of a group after her children would solve this.