move heterogenous child after her parent instead of swapping places
We used to swap places of heteregeneous child and her parent which
moved homogenous parent group higher up in hierarchy. Now the parent
group keeps her place except that an upward child is moved behind
the parent. This may also improve the stability of reordering, and
in the test case I succeeded with one pass instead of two, but still
keeping the repeat{} as it does no harm (unless there are such bad
cases that we go to an infinite loop).