comes empty quadrat, and nothing can be done
Empty quadrats cannot be classified. Empty quadrats could be
handled internally and get class id 0, but supporting function should
be re-written 0-aware, and this is too much toil. Neither we want to
automatically remove empty quadrats (like we do with species) as this
would give a classification vector that does not match data dimensions.
It is better that authors remove empty quadrats so that they know
what was removed.